Tchalla/hamburger component (#99)
* added hamburger component

* added css to spans, created hamburger component

* ran yarn lint to correct lint issues
